feat: create new type Life with relevant resolvers
Introduce new type Life with mandatory fields. Add optional field "title"
to be consistent with previous part. Implement hobbies as an array of
String.

Create mutation to create new lives. Mutation takes in mandatory
information plus optional title argument. Use resolver for mutation.
Resolver concatenates first name and last name to give full name, assigns
a new ObjectID to the new life created, add new life to database and
check for duplicated full name.

Create query and resolver for getLife. Check for match in ID field of
lives. Return the life if there is a match else null.

Create query and resolver for listLives. Get all lives in database and
convert to array. Return array if database is not empty else empty array.

Not sure if there is a need to create the fragment and Apollo Hook in
this part (as per documentation) or if it is only relevant in the next
part.